requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
tqdm>=4.65.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# For data augmentation (optional but recommended)
albumentations>=1.3.0

# For visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# For config management
pyyaml>=6.0